toy_project/monitoring_service/pyproject.toml
[project]
name = "monitoring-service"
version = "0.3.0"
description = "Real-time transaction monitoring microservice with observability"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "pydantic>=2.0",
    "confluent-kafka>=2.6.0",
    "prometheus-client>=0.21.0",
    "opentelemetry-instrumentation-fastapi>=0.46b0",
    "scikit-learn>=1.4.0",
    "joblib>=1.3.0",
    "numpy>=1.26.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
